# web/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lucen AI",
    "description": "Synthetic Identity and Deepfake Insurance-Claim Fraud Detection System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/results/{result_id}": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Analysis Result",
        "operationId": "get_analysis_result_api_v1_results__result_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Results"
        ],
        "summary": "Delete Analysis Result",
        "operationId": "delete_analysis_result_api_v1_results__result_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{result_id}/evidence": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Result Evidence",
        "operationId": "get_result_evidence_api_v1_results__result_id__evidence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          },
          {
            "name": "pipeline",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by pipeline (e.g. image, document, img_1, doc)",
              "title": "Pipeline"
            },
            "description": "Filter by pipeline (e.g. image, document, img_1, doc)"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity (low, medium, high)",
              "title": "Severity"
            },
            "description": "Filter by severity (low, medium, high)"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by kind (risk, authenticity, info)",
              "title": "Kind"
            },
            "description": "Filter by kind (risk, authenticity, info)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "title": "Response Get Result Evidence Api V1 Results  Result Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get History",
        "operationId": "get_history_api_v1_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "band",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{result_id}/report.json": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Json Report",
        "operationId": "get_json_report_api_v1_results__result_id__report_json_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{result_id}/report.pdf": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Pdf Report",
        "operationId": "get_pdf_report_api_v1_results__result_id__report_pdf_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/{result_id}/timeline": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Result Timeline",
        "description": "Returns the chronological evidence timeline (\u00a714.5) with contradictions and undated items.",
        "operationId": "get_result_timeline_api_v1_results__result_id__timeline_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AadhaarQrStatus": {
        "properties": {
          "found": {
            "type": "boolean",
            "title": "Found"
          },
          "signature_valid": {
            "type": "boolean",
            "title": "Signature Valid"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "comparisons": {
            "items": {
              "$ref": "#/components/schemas/QrComparison"
            },
            "type": "array",
            "title": "Comparisons"
          },
          "photo_similarity": {
            "type": "number",
            "title": "Photo Similarity"
          }
        },
        "type": "object",
        "required": [
          "found",
          "signature_valid",
          "mode",
          "comparisons",
          "photo_similarity"
        ],
        "title": "AadhaarQrStatus"
      },
...
      "AnalysisResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "claim_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "image": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "identity": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Identity"
          },
          "claim": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "overall": {
            "$ref": "#/components/schemas/OverallScore"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "detector_status": {
            "items": {
              "$ref": "#/components/schemas/DetectorStatus"
            },
            "type": "array",
            "title": "Detector Status"
          },
          "quality_warnings": {
            "items": {
              "$ref": "#/components/schemas/QualityWarning"
            },
            "type": "array",
            "title": "Quality Warnings"
          },
          "artifacts": {
            "type": "object",
            "title": "Artifacts"
          },
          "versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Versions"
          },
          "location": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "voice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "voice_details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "liveness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LivenessStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "aadhaar_qr": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AadhaarQrStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Link"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Links"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Decision"
              },
              {
                "type": "null"
              }
            ]
          },
          "image_results": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PipelineScore"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Results"
          },
          "why_this_score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WhyThisScore"
              },
              {
                "type": "null"
              }
            ]
          },
          "why_this_score_items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WhyThisScoreItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why This Score Items"
          },
          "checks_run": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CheckRunItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checks Run"
          },
          "top_reasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Reasons"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "summary_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Source",
            "default": "template"
          }
        },
        "type": "object",
        "required": [
          "id",
          "mode",
          "created_at",
          "overall",
          "evidence",
          "detector_status",
          "quality_warnings",
          "artifacts",
          "versions"
        ],
        "title": "AnalysisResult"
      },
...
      "DetectorStatus": {
        "properties": {
          "detector": {
            "type": "string",
            "title": "Detector"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "detector",
          "status",
          "duration_ms"
        ],
        "title": "DetectorStatus"
      },
      "Evidence": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "risk"
          },
          "raw_score": {
            "type": "number",
            "title": "Raw Score",
            "default": 0.5
          },
          "calibrated_score": {
            "type": "number",
            "title": "Calibrated Score",
            "default": 0.5
          },
          "weight": {
            "type": "number",
            "title": "Weight",
            "default": 0.5
          },
          "effective_weight": {
            "type": "number",
            "title": "Effective Weight",
            "default": 0.5
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "medium"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "bboxes": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bboxes"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "artifact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact"
          },
          "pipeline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "contribution": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution"
          },
          "contribution_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contribution Pct"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "pipeline_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Input"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "HistoryResponse"
      },
...
      "LivenessStatus": {
        "properties": {
          "performed": {
            "type": "boolean",
            "title": "Performed"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "code_match": {
            "type": "boolean",
            "title": "Code Match"
          },
          "challenges": {
            "items": {
              "$ref": "#/components/schemas/ChallengeStatus"
            },
            "type": "array",
            "title": "Challenges"
          }
        },
        "type": "object",
        "required": [
          "performed",
          "passed",
          "code_match",
          "challenges"
        ],
        "title": "LivenessStatus"
      },
...
      "OverallScore": {
        "properties": {
          "risk": {
            "type": "number",
            "title": "Risk"
          },
          "band": {
            "type": "string",
            "title": "Band"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "risk",
          "band",
          "confidence",
          "summary"
        ],
        "title": "OverallScore"
      },
...
      "PipelineScore": {
        "properties": {
          "pipeline": {
            "type": "string",
            "title": "Pipeline"
          },
          "risk": {
            "type": "number",
            "title": "Risk"
          },
          "authenticity": {
            "type": "number",
            "title": "Authenticity"
          },
          "band": {
            "type": "string",
            "title": "Band"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "type": "object",
        "required": [
          "pipeline",
          "risk",
          "authenticity",
          "band",
          "confidence",
          "evidence_ids"
        ],
        "title": "PipelineScore"
      },
...
      "QualityWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "QualityWarning"
      },
...
      "VoiceDetails": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "translation_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation En"
          },
          "extracted": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceExtracted"
              },
              {
                "type": "null"
              }
            ]
          },
          "spoof": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceSpoof"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "VoiceDetails"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}